102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,15 +7,15 @@
  */
 int main(void)
 {
-	long long fib1 = 1, fib2 = 2, next_fib;
+	unsigned long fib1 = 1, fib2 = 2, next_fib;
 	int count;
 
-	printf("%lld, %lld", fib1, fib2);
+	printf("%lu, %lu", fib1, fib2);
 
 	for (count = 3; count <= 50; count++)
 	{
 		next_fib = fib1 + fib2;
-		printf(", %lld", next_fib);
+		printf(", %lu", next_fib);
 
 		fib1 = fib2;
 		fib2 = next_fib;
